Update for use with Windows 7
The  lines in the configuration dialog box were removed because they
referred to VB PowerPack, which is not included in VS 2015.
Additionally, disabled the audit verification failure check, as the
audit verification has not been fixed.  Data is still collected,
however.

Also fixed the enabling and disabling of data traces on the graph.