Added logging of all communications with Keithleys
Added the ability to enable logging of the commands/responses to the
Keithley hardware.  Logging is disabled by default when the program is
launched.  Logging can be enabled without the use of a password in the
configuration screen.  Logging has the possibility of causing the
readings to exceed the reading interval, so should be used for
diagnostic purposes only.